// frontend/src/app/api/model/sda.ts
/**
 * CloudPulse AI API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Sda { 
    id: string;
    name: string;
    ownerEmail: string;
    team?: string | null;
    tagValues: { [key: string]: string; };
}